+++ src/jquery.fancytree.persist.js
@@ -223,7 +223,7 @@ $.ui.fancytree.registerExtension({
 		// Bind init-handler to apply cookie state
 		tree.$div.bind("fancytreeinit", function(event){
 			var cookie, dfd, i, keyList, node,
-				prevFocus = $.cookie(local.cookiePrefix + FOCUS); // record this before node.setActive() overrides it;
+				prevFocus = local._data(local.cookiePrefix + FOCUS); // record this before node.setActive() overrides it;
 
 			// tree.debug("document.cookie:", document.cookie);
 
@@ -274,14 +274,23 @@ $.ui.fancytree.registerExtension({
 					if(cookie && (opts.persist.overrideSource || !tree.activeNode)){
 						node = tree.getNodeByKey(cookie);
 						if(node){
-							node.setActive();
+							node.debug("persist: set active", cookie);
+							// We only want to set the focus if the container
+							// had the keyboard focus before
+							node.setActive(true, {noFocus: true});
 						}
 					}
 				}
 				if(local.storeFocus && prevFocus){
 					node = tree.getNodeByKey(prevFocus);
 					if(node){
-						node.setFocus();
+						// node.debug("persist: set focus", cookie);
+						if( tree.options.titlesTabbable ) {
+							$(node.span).find(".fancytree-title").focus();
+						} else {
+							$(tree.$container).focus();
+						}
+						// node.setFocus();
 					}
 				}
 				tree._triggerTreeEvent("restore", null, {});
@@ -322,7 +331,7 @@ $.ui.fancytree.registerExtension({
 		flag = (flag !== false);
 		res = this._superApply(arguments);
 
-		if(flag && local.storeFocus){
+		if( local.storeFocus ) {
 			local._data(local.cookiePrefix + FOCUS, this.focusNode ? this.focusNode.key : null);
 		}
 		return res;