+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.7.1-0 / Unreleased
-  *
+  * [Improved] Bluring the widget will now blur the focused node too.
+  * [Improved] Persitence will only set node focus if widget had focus (otherwise only set active node).
+  * [Improved] Set default focus on first keypress to active node (first node otherwise)
+  * [Added] `_superApply()` for hook handlers.
+  * [Fixed] persistence for focus (when using non-cookie storage)
 
 # 2.7.0 / 2014-12-21
   * [CHANGED] Dropped `fx` option. Use `toggleEffect` instead.