+++ CHANGELOG.md
@@ -2,8 +2,9 @@
   * [Improved] Bluring the widget will now blur the focused node too.
   * [Improved] Persistence will only set node focus if widget had focus (otherwise only activate the node).
   * [Improved] Set default focus on first keypress to active node (first node otherwise)
-  * [Improved] #383 Accept ECMAScript 6 Promise as source
+  * [Improved] #383 Accept [ECMAScript 6 Promise](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_Objects/Promise) as source
   * [Added] `_superApply()` for hook handlers.
+  * [Added] eventToString() supports mouse events
   * [Fixed] persistence for focus (when using non-cookie storage)
 
 # 2.7.0 / 2014-12-21